requirements/dev.txt
-r base.txt


django-extensions==3.2.1
django-debug-toolbar==4.0.0
aiocsv==1.2.4
aiofiles==23.1.0
